Subject: Remove use of cur_term global struct From: Scott Bahling sbahling@suse.com Date: 2015-10-04
With the move to ncurses6, it has been decided by the openSUSE
developer to build the library in threaded mode only. In threaded
mode, the global variables are protected and not exported.

ghc-terminfo was accessing (reading and writing) the cur_term
global so it doesn't work with the openSUSE threaded ncurses6
library. This is an attempt to remove the need to access cur_term
and rely only on the low-level functions.

First, cur_term was being set to a null pointer in order to force
ncurses setupterm() to return new struct instead of a copy of the
current when the terminal name is the same. Actually (and unfortunately)
setupterm() does not return the struct but rather set the global cur_term.
Reading the ncurses code, and the NEWS file, it appears that the
behavior of setupterm() had been changed quite some time ago, and no
longer returns a copy of the struct (internal 'reuse' option) (NEWS 20041127).
The code seems to confirm this, so setting cur_term to a null pointer
shouldn't really be required - we remove poke cur_term!

That's great, but the we still have the problem that setupterm() doesn't
return the terminal struct that is required later to setup the
cleanup routine of the Terminal type (del_curterm cterm). The only
function that appears to return the cur_term pointer is set_curterm()
which has the side effect of overwriting the current pointer in the
process :/  So, we call set_curterm() with a null pointer, retrieving
the former, cur_term pointer for our use in the cterm instance, and
then call set_curterm() again with the former pointer to set it
back again.... whew! Having looked at the ncurses set_curterm() code
it seems it handles a null pointer in the way we would need.

Later in the withCurTerm function we use virtually the same trick
to relieve the need to peek into the cur_term global. I'm not so
sure if this is Haskell correct code though.

Someone with more Haskell programming knowledge than me should review
this patch. It still needs to be tested as well.